// red-server/src/types/docs.ts
// data extracted from property document
export type PropertyDocument = {
  tenants?: string[]
  property?: string
  unitAddress?: string
  email?: string
  alternateEmail?: string
  residentAddress?: string
  residentAddress2?: string
  cellPhone?: string
  alternatePhone?: string
  unit?: string
  lateFeesBaseAmount?: string
  lateFeesGracePeriod?: string
  previousRent?: string
  secDep?: string
  unitType?: string
  unitSize?: string
  unitSf?: string
  rentAmount?: number
  leaseTerm?: string
  leaseFrom?: string
  leaseSignDate?: string
}